# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Too Good To Go API",
        "description": "It is an application that allows us to create and manage users and recipes. All users will be able to search for all recipes, they will be able to create and update their own recipes, they will be able to add and delete favorite recipes and they will also be able to see other users' recipes.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://toogoodtogo.test:8080"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List all available users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Insert a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get by id the information for a specific user,  and also display the information of their recipes.",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jhon Snow"
                    },
                    "email": {
                        "type": "string",
                        "example": "jhon_snow@example.org"
                    },
                    "is_admin": {
                        "type": "number",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-10-23T00:09:16.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-10-23T12:33:45.000000Z"
                    }
                },
                "type": "object",
                "example": {
                    "id": 25,
                    "name": "Jhon Snow",
                    "email": "jhon_snow@example.org",
                    "id_admin": "0",
                    "created_at": "2023-10-23T00:09:16.000000Z",
                    "updated_at": "2023-10-23T00:10:16.000000Z"
                }
            },
            "UserRequest": {
                "title": "UserRequest",
                "description": "UserRequest model",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jhon Snow"
                    },
                    "email": {
                        "type": "string",
                        "example": "jhon_snow@example.org"
                    },
                    "password": {
                        "type": "string",
                        "example": "2_G0od.2-gO"
                    },
                    "is_admin": {
                        "type": "number",
                        "example": "1"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Jhon Snow",
                    "email": "jhon_snow@example.org",
                    "password": "2_G0od.2-gO",
                    "id_admin": "0"
                }
            },
            "ApiResponse": {
                "title": "ApiResponse",
                "description": "ApiResponse model",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "success"
                    },
                    "status_code": {
                        "type": "number",
                        "example": "200"
                    },
                    "errors": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "data": {
                        "type": "array|object",
                        "example": "[]|{}"
                    }
                },
                "type": "object",
                "example": {
                    "message": "success",
                    "status_code": 200,
                    "errors": false,
                    "data": "[]|{}"
                }
            },
            "ApiResponseError": {
                "title": "ApiResponseError",
                "description": "ApiResponseError model",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error"
                    },
                    "status_code": {
                        "type": "number",
                        "example": "400"
                    },
                    "errors": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "data": {
                        "type": "array|object",
                        "example": "[]|{}"
                    }
                },
                "type": "object",
                "example": {
                    "message": "error",
                    "status_code": 400,
                    "errors": true,
                    "data": "[]|{}"
                }
            }
        }
    }
}